requirements.txt
openai
SpeechRecognition
pyaudio
pyyaml